Added resume tab
